LPD-33334 Move setup tests to tests folder, this way they appear in Testray
