+++ modules/test/playwright/tests/setup/page-management-site/site-initializer/fragments/group/apple/fragment.json
