+++ modules/test/playwright/tests/setup/page-management-site/site-initializer/taxonomy-vocabularies/group/animals/dogs.json
